requirements.txt
# Backend Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
plotly>=5.15.0

# Data Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
SQLAlchemy>=2.0.0

# API and Web Scraping
aiohttp>=3.8.0
schedule>=1.2.0

# Mathematical and Statistical
scipy>=1.11.0
statsmodels>=0.14.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment
python-dotenv>=1.0.0

# Machine Learning
joblib>=1.3.0
